declared variables of Savings a/c class

+++ src/main/java/bank/CurrentAccount.java
@@ -3,7 +3,7 @@ package bank;
 public class CurrentAccount {
 
     private String accountNumber;
-    private int balance;
+    private float balance;
     private int withDrawalLimit;
     private int transactionCharges;
     private int insufficentFundCharges;
